Initial Code from 1994
Found tarfile of oldest code & committing here.  Date is from 1993 for
some of this.  A man file output (not included) gives 15 November 1991
as the starting point for this code.  It was called Herbie at that
time.  I forget why it was changed to Izzy.  I think I changed it back
at some point.